🚚 Multi: Move functions around according to their actual use
