+++ utils/utils.go
@@ -15,116 +15,3 @@ limitations under the License.
 */
 
 package utils
-
-import (
-	"errors"
-	"fmt"
-	"io"
-	"os"
-	"regexp"
-	"slices"
-	"strconv"
-	"strings"
-
-	"gopkg.in/yaml.v3"
-)
-
-var (
-	FileOpener = os.Open
-)
-
-// ToPath creates path from path and key name
-func ToPath(path, key string) string {
-	if len(path) == 0 {
-		return key
-	} else {
-		return fmt.Sprintf("%s.%s", path, key)
-	}
-}
-
-// ToListPath like ToPath, but for lists
-func ToListPath(path string, index int) string {
-	sub := fmt.Sprintf("[%d]", index)
-	if len(path) == 0 {
-		return sub
-	} else {
-		return path + sub
-	}
-}
-
-var listPropRe = regexp.MustCompile(".*(\\[\\d+])+")
-
-func ParseListPathComponent(path string) (string, []int, bool) {
-	if !listPropRe.MatchString(path) {
-		return "", nil, false
-	}
-	indexes := make([]int, 0)
-	first := strings.Index(path, "[")
-	cpath := path
-	for {
-		start := strings.Index(cpath, "[")
-		if start == -1 {
-			return path[0:first], indexes, true
-		}
-		end := strings.Index(cpath, "]")
-		index, _ := strconv.Atoi(cpath[start+1 : end])
-		indexes = append(indexes, index)
-		cpath = cpath[end+1:]
-	}
-}
-
-// NewYamlEncoder creates and uniformly configures yaml.Encoder across project
-func NewYamlEncoder(w io.Writer) *yaml.Encoder {
-	enc := yaml.NewEncoder(w)
-	enc.SetIndent(2)
-	return enc
-}
-
-type failingReader struct{}
-
-func (fr failingReader) Read([]byte) (int, error) {
-	return 0, errors.New("read just failed")
-}
-
-func FailingReader() io.Reader {
-	return &failingReader{}
-}
-
-type failingWriter struct{}
-
-func (fw failingWriter) Write([]byte) (int, error) {
-	return 0, errors.New("write just failed")
-}
-
-func FailingWriter() io.Writer {
-	return &failingWriter{}
-}
-
-func Unique(in []string) []string {
-	ret := make([]string, 0)
-	for _, s := range in {
-		if !slices.Contains(ret, s) {
-			ret = append(ret, s)
-		}
-	}
-	return ret
-}
-
-// Unflatten map entries into new map.
-func Unflatten(in map[string]interface{}) map[string]interface{} {
-	res := make(map[string]interface{})
-	for k, v := range in {
-		current := res
-		pc := strings.Split(k, ".")
-		for _, c := range pc[0 : len(pc)-1] {
-			if x, exists := current[c].(map[string]interface{}); exists {
-				current = x
-			} else {
-				current[c] = make(map[string]interface{})
-				current = current[c].(map[string]interface{})
-			}
-		}
-		current[pc[len(pc)-1]] = v
-	}
-	return res
-}
