+++ utils/utils_test.go
@@ -15,65 +15,3 @@ limitations under the License.
 */
 
 package utils
-
-import (
-	"bytes"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestToPath(t *testing.T) {
-	assert.Equal(t, "abc", ToPath("", "abc"))
-	assert.Equal(t, "abc.def", ToPath("abc", "def"))
-}
-
-func TestNewYamlEncoder(t *testing.T) {
-	assert.NotNil(t, NewYamlEncoder(bytes.NewBuffer(make([]byte, 0))))
-}
-
-func TestParseListPathComponent(t *testing.T) {
-	name, indexes, ok := ParseListPathComponent("list[0][3][1]")
-	assert.True(t, ok)
-	assert.Equal(t, 3, len(indexes))
-	assert.Equal(t, 0, indexes[0])
-	assert.Equal(t, 3, indexes[1])
-	assert.Equal(t, 1, indexes[2])
-	assert.Equal(t, "list", name)
-	_, _, ok = ParseListPathComponent("not a list")
-	assert.False(t, ok)
-}
-
-func TestToListPath(t *testing.T) {
-	assert.Equal(t, "[2]", ToListPath("", 2))
-	assert.Equal(t, "item[3]", ToListPath("item", 3))
-}
-
-func TestFailingReader(t *testing.T) {
-	_, err := FailingReader().Read([]byte{})
-	assert.Error(t, err)
-}
-
-func TestFailingWriter(t *testing.T) {
-	_, err := FailingWriter().Write([]byte{})
-	assert.Error(t, err)
-}
-
-func TestUnique(t *testing.T) {
-	assert.Equal(t, 3, len(Unique([]string{"a", "a", "b", "x", "x"})))
-}
-
-func TestUnflattenMap(t *testing.T) {
-	out := Unflatten(map[string]interface{}{
-		"a.b": 123,
-		"a.c": "0876",
-		"x.y": "hello",
-		"x.z": nil,
-		"s":   3.14,
-	})
-	assert.Equal(t, 123, out["a"].(map[string]interface{})["b"])
-	assert.Equal(t, "0876", out["a"].(map[string]interface{})["c"])
-	assert.Equal(t, "hello", out["x"].(map[string]interface{})["y"])
-	assert.Equal(t, nil, out["x"].(map[string]interface{})["z"])
-	assert.Equal(t, 3.14, out["s"])
-}
